requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pymongo==4.6.0
firecrawl-py==0.0.16
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
